% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{caching}
\alias{caching}
\title{Create a function that honours HTTP caching headers}
\usage{
caching(f, cache = cachem::cache_mem())
}
\arguments{
\item{f}{A function that returns an \link[httr:response]{httr::response}.}

\item{cache}{A cache object to control caching. See \code{\link[cachem:cache_mem]{cachem::cache_mem()}}.}
}
\value{
\code{f} with the response cached according to \code{\link[httr:cache_info]{httr::cache_info()}}. The
given \code{cache} object is available in the \code{"cache"} attribute.
}
\description{
Create a function that honours HTTP caching headers
}
\examples{
# Use a disk cache for persistent caching across sessions
disk_cached_get <- caching(httr::GET, cachem::cache_disk())

# Inspect the cache object
str(attr(disk_cached_get, "cache")$info())
}
\seealso{
Pre-wrapped \link{httr} \link{verbs}.
}
