% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httc-package.R
\docType{package}
\name{httc-package}
\alias{httc}
\alias{httc-package}
\title{httc: HTTP Requests Honouring Cache Headers}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikmart@pm.me}

}
\keyword{internal}
