% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{verbs}
\alias{verbs}
\alias{BROWSE}
\alias{DELETE}
\alias{GET}
\alias{HEAD}
\alias{PATCH}
\alias{POST}
\alias{PUT}
\alias{VERB}
\title{Caching versions of httr verbs}
\usage{
BROWSE(...)

DELETE(...)

GET(...)

HEAD(...)

PATCH(...)

POST(...)

PUT(...)

VERB(...)
}
\arguments{
\item{...}{Arguments passed on to the corresponding \link{httr} verb.}
}
\value{
Result of original verb, but from a cache if applicable.
}
\description{
Memory-caching versions of \link{httr} verbs that honour cache control headers. If
you need more control over the caching, create your own with \code{\link[=caching]{caching()}}.
}
\examples{
# Response header specifies two seconds of caching
cached_request <- function() {
  try(httc::GET("https://httpbin.org/cache/2"))
}

system.time(cached_request()) # First response from server
system.time(cached_request()) # Second uses cached result
Sys.sleep(2)
system.time(cached_request()) # Re-requested from server

}
